Ensure user documents exist before creating profiles
