+++ app/api/avatar/route.ts
@@ -1,14 +1,7 @@
 import { NextResponse, type NextRequest } from "next/server";
 import { currentUser } from "@clerk/nextjs/server";
-import { createClient } from "next-sanity";
-
-const client = createClient({
-  projectId: "8n5iznjt",
-  dataset: "production",
-  apiVersion: "2025-06-09",
-  token: process.env.SANITY_API_TOKEN,
-  useCdn: false,
-});
+import { client } from "@/app/sanity/client";
+import { ensureUser } from "@/app/sanity/user";
 
 export async function POST(req: NextRequest) {
   const user = await currentUser();
@@ -29,6 +22,11 @@ export async function POST(req: NextRequest) {
     { id: user.id }
   );
   if (!profileId) {
+    await ensureUser({
+      id: user.id,
+      email: user.primaryEmailAddress?.emailAddress,
+      fullName: user.fullName,
+    });
     const created = await client.create({
       _type: "profile",
       user: { _type: "reference", _ref: user.id },
