+++ app/api/profile/route.ts
@@ -1,14 +1,7 @@
 import { NextResponse, type NextRequest } from "next/server";
 import { currentUser } from "@clerk/nextjs/server";
-import { createClient } from "next-sanity";
-
-const client = createClient({
-  projectId: "8n5iznjt",
-  dataset: "production",
-  apiVersion: "2025-06-09",
-  token: process.env.SANITY_API_TOKEN,
-  useCdn: false,
-});
+import { client } from "@/app/sanity/client";
+import { ensureUser } from "@/app/sanity/user";
 
 export async function POST(req: NextRequest) {
   const user = await currentUser();
@@ -18,14 +11,11 @@ export async function POST(req: NextRequest) {
 
   const data = await req.json();
 
-  const userDoc = {
-    _type: "user",
-    _id: user.id,
-    clerkId: user.id,
+  await ensureUser({
+    id: user.id,
     email: user.primaryEmailAddress?.emailAddress,
     fullName: data.fullName,
-    role: "user",
-  };
+  });
 
   const profileDoc = {
     _type: "profile",
@@ -38,7 +28,6 @@ export async function POST(req: NextRequest) {
     location: data.location,
   };
 
-  await client.createIfNotExists(userDoc);
   const created = await client.createIfNotExists(profileDoc);
 
   return NextResponse.json({ profile: created });
@@ -56,6 +45,11 @@ export async function GET() {
   );
 
   if (!profile) {
+    await ensureUser({
+      id: user.id,
+      email: user.primaryEmailAddress?.emailAddress,
+      fullName: user.fullName,
+    });
     profile = await client.create({
       _type: "profile",
       user: { _type: "reference", _ref: user.id },
@@ -78,6 +72,11 @@ export async function PUT(req: NextRequest) {
   );
 
   if (!profileId) {
+    await ensureUser({
+      id: user.id,
+      email: user.primaryEmailAddress?.emailAddress,
+      fullName: user.fullName,
+    });
     const created = await client.create({
       _type: "profile",
       user: { _type: "reference", _ref: user.id },
